+++ app/src/main/kotlin/com/tajmoti/tulip/ui/player/episodes/AndroidEpisodesViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.tajmoti.libtulip.model.key.TvShowKey
 import com.tajmoti.libtulip.repository.FavoritesRepository
 import com.tajmoti.libtulip.repository.HostedTvDataRepository
+import com.tajmoti.libtulip.repository.PlayingHistoryRepository
 import com.tajmoti.libtulip.repository.TmdbTvDataRepository
 import com.tajmoti.libtulip.ui.tvshow.TvShowViewModel
 import com.tajmoti.libtulip.ui.tvshow.TvShowViewModelImpl
@@ -17,12 +18,14 @@ class AndroidEpisodesViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
     hostedTvDataRepository: HostedTvDataRepository,
     tmdbRepo: TmdbTvDataRepository,
-    favoritesRepository: FavoritesRepository
+    favoritesRepository: FavoritesRepository,
+    historyRepository: PlayingHistoryRepository
 ) : DelegatingViewModel<TvShowViewModel>() {
     override val impl = TvShowViewModelImpl(
         hostedTvDataRepository,
         tmdbRepo,
         favoritesRepository,
+        historyRepository,
         viewModelScope,
         savedStateHandle.get<TvShowKey>(EpisodesFragment.ARG_TV_SHOW_KEY)!!
     )