+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Changed
 - Transparent status bar on Library screen
+- Preselect last played seasons when enterin TV show screen
 
 ## [0.12.0] - 2021-11-07
 ### Added